fixed IDs format issue, fixed/implemented update option on crawler and parser and changed files directory
IDs are inverted and pronouns (no(s)/na(s)) are excluded. ex: AI no REsp na AgRg 1234 ->  AgRg 1234 REsp AI

To run type on the terminal:
	scrapy crawl -a "iDate"="20010101" -a "fDate"="20030101" -a "update"=0 stj_acordaos

setting iDate as the initial date of the search and fDate the final date in the format yyyymmdd
to update set arg "update"=1 and be sure to set the fDate
updating will take the next day of the previous search fDate (registered on the file update_settings on the root folder of the project)
to be the initial date of search.
If not updating, the search parameters will overwrite the ones in the update settings file and the files will be indexed from #1

Files will be on the format stj######.xml on "files" folder on the project root