added methods to test ids, ufs and relators
method writes to files, list of ids, relators and ufs found

Already found a bug on relators as unicode characters aren't being captured.
solution could be not to use 'text()' xpath attribute on extracting this info,
but then can't use startswith(), should use search regex method

tried to parse the result, which is inside observation section, but it is hard
as there are multiple forms of ending.